+++ tiger-test-lib/src/main/java/de/gematik/test/tiger/lib/TigerLibConfig.java
@@ -40,4 +40,5 @@ public class TigerLibConfig {
   @Builder.Default public TigerHttpClientConfig httpClientConfig = new TigerHttpClientConfig();
   @Builder.Default private boolean trafficVisualization = false;
   @Builder.Default public boolean clearEnvironmentStartupTraffic = true;
+  @Builder.Default public boolean runTestsOnStart = true;
 }