+++ tiger-uitests/src/test/java/de/gematik/test/tiger/playwright/workflowui/StaticRbelLogTests.java
@@ -15,10 +15,9 @@
  */
 
 package de.gematik.test.tiger.playwright.workflowui;
-
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 
+import com.microsoft.playwright.assertions.PlaywrightAssertions;
 import org.junit.jupiter.api.Test;
 
 /** Tests for static content of the web ui content, e.g. rbel logo. */
@@ -28,7 +27,7 @@ class StaticRbelLogTests extends AbstractTests {
   void testExecutionPaneRbelLogo() {
     page.querySelector("#test-execution-pane-tab").click();
     page.locator("#test-webui-slider").click();
-    await().untilAsserted(() -> assertThat(page.locator("#test-rbel-logo").isVisible()).isTrue());
+    await().untilAsserted(() -> PlaywrightAssertions.assertThat(page.locator("#test-rbel-logo")).isVisible());
     screenshot(page, "maincontent_rbelpath.png");
     screenshotElementById(
         page, "maincontent_rbelpath_urllink_highlight.png", "test-rbel-webui-url");