+++ tiger-uitests/src/test/java/de/gematik/test/tiger/playwright/workflowui/StaticSidebarTests.java
@@ -16,7 +16,7 @@
 
 package de.gematik.test.tiger.playwright.workflowui;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import lombok.extern.slf4j.Slf4j;
@@ -29,49 +29,47 @@ class StaticSidebarTests extends AbstractTests {
   @Test
   void testSidebarClosedIconsAreVisible() {
     assertAll(
-        () -> assertThat(page.querySelector("#test-sidebar-title").isVisible()).isFalse(),
-        () -> assertThat(page.querySelector("div.test-sidebar-collapsed").isVisible()).isTrue(),
-        () -> assertThat(page.querySelector("#test-tiger-logo").isVisible()).isTrue(),
-        () -> assertThat(page.querySelector("#test-sidebar-quit-icon").isVisible()).isTrue(),
-        () -> assertThat(page.querySelector("#test-sidebar-pause-icon").isVisible()).isTrue(),
-        () -> assertThat(page.querySelector("#test-sidebar-status-icon").isVisible()).isTrue(),
-        () -> assertThat(page.querySelector("#test-sidebar-feature-icon").isVisible()).isTrue(),
-        () -> assertThat(page.querySelector("#test-sidebar-server-icon").isVisible()).isTrue());
+        () -> assertThat(page.locator("#test-sidebar-title")).not().isVisible(),
+        () -> assertThat(page.locator("div.test-sidebar-collapsed")).isVisible(),
+        () -> assertThat(page.locator("#test-tiger-logo")).isVisible(),
+        () -> assertThat(page.locator("#test-sidebar-quit-icon")).isVisible(),
+        () -> assertThat(page.locator("#test-sidebar-pause-icon")).isVisible(),
+        () -> assertThat(page.locator("#test-sidebar-status-icon")).isVisible(),
+        () -> assertThat(page.locator("#test-sidebar-feature-icon")).isVisible(),
+        () -> assertThat(page.locator("#test-sidebar-server-icon")).isVisible());
   }
 
   @Test
   void testSidebarOpenIconsAreVisible() {
     page.querySelector("#test-tiger-logo").click();
     assertAll(
-        () -> assertThat(page.querySelector("#test-sidebar-title").isVisible()).isTrue(),
-        () -> assertThat(page.querySelector("#test-sidebar-status").isVisible()).isTrue(),
-        () -> assertThat(page.querySelector("#test-sidebar-feature").isVisible()).isTrue(),
-        () -> assertThat(page.querySelector("#test-sidebar-server").isVisible()).isTrue(),
-        () -> assertThat(page.querySelector("#test-sidebar-version").isVisible()).isTrue(),
-        () -> assertThat(page.querySelector("#test-sidebar-build").isVisible()).isTrue(),
-        () -> assertThat(page.querySelector("#test-tiger-logo").isVisible()).isTrue(),
-        () -> assertThat(page.querySelector("#test-sidebar-quit-icon").isVisible()).isTrue(),
-        () -> assertThat(page.querySelector("#test-sidebar-pause-icon").isVisible()).isTrue(),
-        () -> assertThat(page.querySelector("#test-sidebar-status-icon").isVisible()).isTrue(),
-        () -> assertThat(page.querySelector("#test-sidebar-feature-icon").isVisible()).isTrue(),
-        () -> assertThat(page.querySelector("#test-sidebar-server-icon").isVisible()).isTrue());
+        () -> assertThat(page.locator("#test-sidebar-title")).isVisible(),
+        () -> assertThat(page.locator("#test-sidebar-status")).isVisible(),
+        () -> assertThat(page.locator("#test-sidebar-feature")).isVisible(),
+        () -> assertThat(page.locator("#test-sidebar-server")).isVisible(),
+        () -> assertThat(page.locator("#test-sidebar-version")).isVisible(),
+        () -> assertThat(page.locator("#test-sidebar-build")).isVisible(),
+        () -> assertThat(page.locator("#test-tiger-logo")).isVisible(),
+        () -> assertThat(page.locator("#test-sidebar-quit-icon")).isVisible(),
+        () -> assertThat(page.locator("#test-sidebar-pause-icon")).isVisible(),
+        () -> assertThat(page.locator("#test-sidebar-status-icon")).isVisible(),
+        () -> assertThat(page.locator("#test-sidebar-feature-icon")).isVisible(),
+        () -> assertThat(page.locator("#test-sidebar-server-icon")).isVisible());
   }
 
   @Test
   void testStatus() {
     page.querySelector("#test-tiger-logo").click();
     assertAll(
-        () -> assertThat(page.locator("#test-sidebar-statusbox").isVisible()).isTrue(),
+        () -> assertThat(page.locator("#test-sidebar-statusbox")).isVisible(),
         () ->
             assertThat(
-                    page.locator("#test-sidebar-statusbox .test-sidebar-status-features")
-                        .isVisible())
-                .isTrue(),
+                    page.locator("#test-sidebar-statusbox .test-sidebar-status-features"))
+                        .isVisible(),
         () ->
             assertThat(
-                    page.locator("#test-sidebar-statusbox .test-sidebar-status-scenarios")
-                        .isVisible())
-                .isTrue(),
-        () -> assertThat(page.locator("#test-sidebar-status-started").isVisible()).isTrue());
+                    page.locator("#test-sidebar-statusbox .test-sidebar-status-scenarios"))
+                        .isVisible(),
+        () -> assertThat(page.locator("#test-sidebar-status-started")).isVisible());
   }
 }